requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain-google-genai==1.0.2
langchain-community==0.2.0
langchain-core==0.1.60
pydantic==2.6.4
python-dotenv==1.0.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
